refactor(connections): extract CatalogItemCard into its own file
connections.tsx is a 1700+ line God-component. CatalogItemCard only reads
its own props and local state, so it moves out byte-identical.
